requirements.txt
numpy>=1.16.0
pandas>=0.24.0
urllib3>=1.24.0
scikit-learn>=0.20.0
torch==1.6.0
dgl-cu101==0.5.3
tensorflow==2.2.0
